Implement more flexible scheme for GPU strings in ELPA
ELPA uses the idiom

    'elpa_obj%set(gpu_string,1)'

to enable a specific GPU type (e.g. elpa_obj%set('nvidia-gpu',1) ).

Different versions of ELPA support different gpu_string's, which is
a source of confusion and bugs. For example, the 2020.05.001 version
does not support 'nvidia-gpu', whereas later versions consider the
older 'gpu' string as deprecated.

Rather than hardwiring a specific gpu_string, it is better to be
able to set it depending on the ELPA version being used.

For "built-in" versions of ELPA in ELSI, this is achieved directly
by setting the CMake variable ELSI_ELPA_GPU_STRING in the CMakeLists.txt
file in external/ELSI. This variable is used during the generation
of the 'elsi_version.f90' file, in a new subroutine that is called
during ELPA configuration.

For external versions of ELPA, it is the responsibility of the user
to make sure that the variable is set. By default it is set to 'gpu'.

It might be possible to discover the right gpu_string in an external
library, but this is left for the future.